Get A Fresh API Instance For Each Sync Or Download
Get a fresh instance of the ComicLib-API, -Comics or -Images for each
sync or download. Necessary to avoid syncing to the old address after
switching the server address (e.g. in the case of server migration).
Previously, an app restart would have been necessary to start using the
new server address.

Signed-off-by: ahahn94 <ahahn94@outlook.com>